pyproject.toml
[project]
name = "career-conversations-ai"
version = "0.1.0"
description = "RAG-based career assistant built with FastAPI, Qdrant, Redis, and Gradio."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.6.0",
    "openai>=2.21.0",
    "pypdf>=6.7.2",
    "pydantic-settings>=2.6.0",
    "requests>=2.32.5",
    "qdrant-client>=1.12.0",
    "fastapi[standard-no-fastapi-cloud-cli]>=0.131.0",
    "upstash-redis>=1.6.0",
    "loguru>=0.7.3",
]
